Copied classes
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@4756 06c8f799-c779-4783-8af7-179afedb67c8

+++ pccomms/lejos/robotics/objectdetection/TouchFeatureDetector.java
@@ -23,11 +23,27 @@ public class TouchFeatureDetector extends FeatureDetectorAdapter {
 	
 	private ArrayList<FeatureListener> listeners = null;
 	
+	/**
+	 * Creates a touch detector in which the touch sensor is assumed to be situated in the center of
+	 * the robot. This is fine in situations where the robot only needs to react to a bumper
+	 * contact. See the alternate constructor if more than one bumper is located around the robot. 
+	 * @param touchSensor The touch sensor bumper.
+	 */
 	public TouchFeatureDetector(Touch touchSensor) {
 		this(touchSensor, 0, 0);
 	}
 	
-	public TouchFeatureDetector(Touch touchSensor, int xOffset, int yOffset) {
+	/**
+	 * If you want the bumpers to report contact relative to the geometry of where they are placed on the robot,
+	 * you can provide the x, y offsets of each bumper relative to the center of the robot (the center is
+	 * the halfway point between the drive wheels). Most bumpers are planar, so generally you would use the center of
+	 * the bumper as the contact point.
+	 *  
+	 * @param touchSensor The touch sensor bumper.
+	 * @param xOffset The offset (in units e.g. cm) left or right of center. Right is positive, left is negative.
+	 * @param yOffset The offset (in units e.g. cm) forward or back of center. Forward is positive, back is negative.
+	 */
+	public TouchFeatureDetector(Touch touchSensor, double xOffset, double yOffset) {
 		super(DELAY);
 		this.touch_sensor = touchSensor;
 		//this.x_offset = xOffset;
@@ -35,7 +51,7 @@ public class TouchFeatureDetector extends FeatureDetectorAdapter {
 		
 		// Calculate angle a distance of bumper from center:
 		Point robot_center = new Point(0, 0);
-		Point bumper_p = new Point(xOffset, yOffset);
+		Point bumper_p = new Point((float)xOffset, (float)yOffset);
 		range = (float)robot_center.distance(xOffset, yOffset);
 		angle = robot_center.angleTo(bumper_p) - 90;
 	}
